fixing snakemake issues

+++ amst_utils/common/settings.py
@@ -20,8 +20,17 @@ def get_run_info():
 
 
 def set_run_info(run_info):
-    with open(get_run_info_fp(), mode='w') as f:
+    fp = get_run_info_fp()
+    if os.path.exists(fp):
+        raise RuntimeError(f'Another instance of the AMST pre-alignment is already running! '
+                           f'If this is not the case delete {fp} and start again.')
+    with open(fp, mode='w') as f:
         json.dump(run_info, f, indent=2)
+
+
+def remove_run_info():
+    fp = get_run_info_fp()
+    os.remove(fp)
 
 
 def get_params_fp(folder=None):
